backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.8.3
pypdf==5.1.0
pydantic==2.5.3
sentence-transformers==3.3.1
numpy>=1.26.0
scikit-learn>=1.3.2
rank-bm25==0.2.2
